collapse all repository code in one class
