+++ src/Tribe/Repository/Read_Interface.php
@@ -5,11 +5,7 @@
  *
  * @since TBD
  */
-interface Tribe__Repository__Read_Interface
-	extends Tribe__Repository__Specialized_Repository_Interface {
-	const PERMISSION_EDITABLE = 'editable';
-	const PERMISSION_READABLE = 'readable';
-
+interface Tribe__Repository__Read_Interface {
 	/**
 	 * Batch filter application method.
 	 *
@@ -18,10 +14,7 @@ interface Tribe__Repository__Read_Interface
 	 * @since TBD
 	 *
 	 * @param array $args An associative array of arguments to filter
-	 *                    the posts by in the shape [ <key>, <value> ].
-	 *
-	 * @return Tribe__Repository__Read_Interface|Tribe__Repository__Update_Interface
-	 */
+	 *                    the posts by in the shape [ <key>, <value> ]. * * @return Tribe__Repository__Read_Interface */
 	public function by_args( array $args );
 
 	/**
@@ -29,14 +22,7 @@ interface Tribe__Repository__Read_Interface
 	 *
 	 * This is the same as calling `where` multiple times with different arguments.
 	 *
-	 * @since TBD
-	 *
-	 * @param array $args An associative array of arguments to filter
-	 *                    the posts by in the shape [ <key>, <value> ].
-	 *
-	 * @return Tribe__Repository__Read_Interface|Tribe__Repository__Update_Interface
-	 */
-	public function where_args( array $args );
+	 * T
 
 	/**
 	 * Applies a filter to the query.
@@ -51,7 +37,7 @@ interface Tribe__Repository__Read_Interface
 	 * @param mixed  ...$args Additional, optional, call arguments that will be passed to
 	 *                        the schema.
 	 *
-	 * @return Tribe__Repository__Read_Interface|Tribe__Repository__Update_Interface
+	 * @return Tribe__Repository__Read_Interface
 	 */
 	public function by( $key, $value );
 
@@ -63,7 +49,7 @@ interface Tribe__Repository__Read_Interface
 	 * @param string $key
 	 * @param mixed  $value
 	 *
-	 * @return Tribe__Repository__Read_Interface|Tribe__Repository__Update_Interface
+	 * @return Tribe__Repository__Read_Interface
 	 */
 	public function where( $key, $value );
 
@@ -77,7 +63,7 @@ interface Tribe__Repository__Read_Interface
 	 *
 	 * @param int $page
 	 *
-	 * @return Tribe__Repository__Read_Interface|Tribe__Repository__Update_Interface
+	 * @return Tribe__Repository__Read_Interface
 	 */
 	public function page( $page );
 
@@ -89,7 +75,7 @@ interface Tribe__Repository__Read_Interface
 	 *
 	 * @param int $per_page
 	 *
-	 * @return Tribe__Repository__Read_Interface|Tribe__Repository__Update_Interface
+	 * @return Tribe__Repository__Read_Interface
 	 */
 	public function per_page( $per_page );
 
@@ -126,7 +112,7 @@ interface Tribe__Repository__Read_Interface
 	 * @param bool $increment Whether to increment the offset by the value
 	 *                        or replace it.
 	 *
-	 * @return Tribe__Repository__Read_Interface|Tribe__Repository__Update_Interface
+	 * @return Tribe__Repository__Read_Interface
 	 */
 	public function offset( $offset, $increment = false );
 
@@ -140,7 +126,7 @@ interface Tribe__Repository__Read_Interface
 	 *
 	 * @param string $order
 	 *
-	 * @return Tribe__Repository__Read_Interface|Tribe__Repository__Update_Interface
+	 * @return Tribe__Repository__Read_Interface
 	 */
 	public function order( $order = 'ASC' );
 
@@ -154,7 +140,7 @@ interface Tribe__Repository__Read_Interface
 	 *
 	 * @param string $order_by
 	 *
-	 * @return Tribe__Repository__Read_Interface|Tribe__Repository__Update_Interface
+	 * @return Tribe__Repository__Read_Interface
 	 */
 	public function order_by( $order_by );
 
@@ -168,7 +154,7 @@ interface Tribe__Repository__Read_Interface
 	 *
 	 * @param string $fields
 	 *
-	 * @return Tribe__Repository__Read_Interface|Tribe__Repository__Update_Interface
+	 * @return Tribe__Repository__Read_Interface
 	 */
 	public function fields( $fields );
 
@@ -180,7 +166,7 @@ interface Tribe__Repository__Read_Interface
 	 *
 	 * @param string $permission One of the two `self::PERMISSION` constants.
 	 *
-	 * @return Tribe__Repository__Read_Interface|Tribe__Repository__Update_Interface
+	 * @return Tribe__Repository__Read_Interface
 	 */
 	public function permission( $permission );
 
@@ -193,7 +179,7 @@ interface Tribe__Repository__Read_Interface
 	 *
 	 * @param array|int $post_ids
 	 *
-	 * @return Tribe__Repository__Read_Interface|Tribe__Repository__Update_Interface
+	 * @return Tribe__Repository__Read_Interface
 	 */
 	public function in( $post_ids );
 
@@ -206,7 +192,7 @@ interface Tribe__Repository__Read_Interface
 	 *
 	 * @param array|int $post_ids
 	 *
-	 * @return Tribe__Repository__Read_Interface|Tribe__Repository__Update_Interface
+	 * @return Tribe__Repository__Read_Interface
 	 */
 	public function not_in( $post_ids );
 
@@ -219,7 +205,7 @@ interface Tribe__Repository__Read_Interface
 	 *
 	 * @param array|int $post_id
 	 *
-	 * @return Tribe__Repository__Read_Interface|Tribe__Repository__Update_Interface
+	 * @return Tribe__Repository__Read_Interface
 	 */
 	public function parent( $post_id );
 
@@ -232,7 +218,7 @@ interface Tribe__Repository__Read_Interface
 	 *
 	 * @param array $post_ids
 	 *
-	 * @return Tribe__Repository__Read_Interface|Tribe__Repository__Update_Interface
+	 * @return Tribe__Repository__Read_Interface
 	 */
 	public function parent_in( $post_ids );
 
@@ -245,7 +231,7 @@ interface Tribe__Repository__Read_Interface
 	 *
 	 * @param array $post_ids
 	 *
-	 * @return Tribe__Repository__Read_Interface|Tribe__Repository__Update_Interface
+	 * @return Tribe__Repository__Read_Interface
 	 */
 	public function parent_not_in( $post_ids );
 
@@ -258,7 +244,7 @@ interface Tribe__Repository__Read_Interface
 	 *
 	 * @param $search
 	 *
-	 * @return Tribe__Repository__Read_Interface|Tribe__Repository__Update_Interface
+	 * @return Tribe__Repository__Read_Interface
 	 */
 	public function search( $search );
 
@@ -278,16 +264,6 @@ interface Tribe__Repository__Read_Interface
 	public function count();
 
 	/**
-	 * Sets the dynamic part of the filter tag that will be used to filter
-	 * the query arguments and object.
-	 *
-	 * @param string $filter_name
-	 *
-	 * @return Tribe__Repository__Read_Interface|Tribe__Repository__Update_Interface
-	 */
-	public function filter_name( $filter_name );
-
-	/**
 	 * Returns the first post of the page matching the current query.
 	 *
 	 * If, by default or because set with the `per_page` method, all
@@ -298,7 +274,7 @@ interface Tribe__Repository__Read_Interface
 	 *
 	 * @return WP_Post|mixed|null
 	 *
-	 * @see Tribe__Repository__Read_Interface::per_page()
+	 * @see   Tribe__Repository__Read_Interface::per_page()
 	 */
 	public function first();
 
@@ -313,7 +289,7 @@ interface Tribe__Repository__Read_Interface
 	 *
 	 * @return WP_Post|mixed|null
 	 *
-	 * @see Tribe__Repository__Read_Interface::per_page()
+	 * @see   Tribe__Repository__Read_Interface::per_page()
 	 */
 	public function last();
 
@@ -346,7 +322,7 @@ interface Tribe__Repository__Read_Interface
 	 *
 	 * @return array An array of posts matching the query.
 	 *
-	 * @see Tribe__Repository__Read_Interface::per_page()
+	 * @see   Tribe__Repository__Read_Interface::per_page()
 	 */
 	public function take( $n );
 
@@ -382,5 +358,5 @@ interface Tribe__Repository__Read_Interface
 	 *
 	 * @return mixed
 	 */
-	public function get_query(  );
+	public function get_query();
 }
